<feat> 게시글 삭제시 작성자 본인인지 확인 검증, 게시글 좋아요 클릭시 로그인 검증 및 게시글 존재 여부 확인 로직 구현

+++ src/main/java/highsquare/hirecoder/web/controller/BoardController.java
@@ -1,5 +1,7 @@
 package highsquare.hirecoder.web.controller;
 
+import highsquare.hirecoder.constant.SessionConstant;
+import highsquare.hirecoder.domain.repository.BoardRepository;
 import highsquare.hirecoder.domain.repository.StudyMemberRepository;
 import highsquare.hirecoder.domain.service.*;
 import highsquare.hirecoder.entity.*;
@@ -20,7 +22,9 @@ import javax.servlet.http.HttpSession;
 import java.io.IOException;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import static highsquare.hirecoder.constant.PageConstant.*;
 
@@ -38,6 +42,8 @@ public class BoardController {
 
     private final StudyMemberRepository studyMemberRepository;
 
+    private final BoardRepository boardRepository;
+
     private final TagService tagService;
 
 
@@ -50,7 +56,7 @@ public class BoardController {
         HttpSession session = request.getSession(true);
         // 현재 로그인에 관한 페이지가 없으므로 session 값에 강욱 멤버의 memberId를 임의로 넣어두겠음
         // 현재 로그인한 멤버의 name도 넣어두겠음
-        session.setAttribute("memberId",1L);
+        session.setAttribute(SessionConstant.MEMBER_ID,1L);
         session.setAttribute("memberName", "강욱");
 
 
@@ -67,7 +73,7 @@ public class BoardController {
 
         //전체 공개 여부에 따라 멤버가 읽을 수 있는지 없는지 여부
         if (!boardService.isPublic(boardId)) {
-            Long memberId = (Long) session.getAttribute("memberId");
+            Long memberId = (Long) session.getAttribute(SessionConstant.MEMBER_ID);
 
             if(!studyMemberRepository.existsMemberAndStudy(studyId, memberId)) {
                 ScriptUtils.alertAndBackPage(response,"해당 스터디에 해당되지 않습니다.");
@@ -80,7 +86,7 @@ public class BoardController {
 
 
         // 세션에 저장된 member와 해당 게시글을 이용해서 LikeOnBoard 엔티티 가져오기
-        LikeOnBoard likeOnBoard = likeOnBoardService.getLikeOnBoard(board.getId(), (Long)session.getAttribute("memberId"));
+        LikeOnBoard likeOnBoard = likeOnBoardService.getLikeOnBoard(board.getId(), (Long)session.getAttribute(SessionConstant.MEMBER_ID));
 
         // board 엔티티를 boardForm으로 변환
         BoardSelectedForm boardForm = turnBoardEntityToForm(board);
@@ -97,12 +103,12 @@ public class BoardController {
         // DB에서 board.id에 해당하는 PageResultDto<CommentSelectedForm>를 꺼내옴
         PageResultDto<CommentSelectedForm, Comment> allComments =
                 commentService.pagingAllComments(boardId,
-                        (Long)session.getAttribute("memberId"),pageRequestDto);
+                        (Long)session.getAttribute(SessionConstant.MEMBER_ID),pageRequestDto);
 
         // DB에서 board.id에 해당하는 Best 댓글순으로 꺼내옴
         PageResultDto<CommentSelectedForm, Comment> bestComments =
                 commentService.pagingBestComments(boardId,
-                        (Long)session.getAttribute("memberId"),pageRequestDto);
+                        (Long)session.getAttribute(SessionConstant.MEMBER_ID),pageRequestDto);
 
         //게시글에 해당하는 총 댓글수 체크
         Integer commentsTotalCounts = commentService.countComments(boardId);
@@ -125,10 +131,20 @@ public class BoardController {
     // 삭제하기를 누르면 해당 게시글을 삭제하고 그 게시글이 속했던 스터디 게시글 페이지로 이동하게 했음
     // 삭제 시 foreignKey 제약조건으로 인해 삭제가 안됨 -- 설정해야함
 
-    @GetMapping("/content/{study_id}/{board_id}/delete")
+    @GetMapping("/{kind}/{study_id}/{board_id}/delete")
     public String deleteBoard(@PathVariable("study_id") Long studyId,
-                              @PathVariable("board_id") Long boardId, RedirectAttributes redirectAttributes) {
-        boardService.deleteBoard(boardId);
+                              @PathVariable("board_id") Long boardId,
+                              HttpSession session,
+                              RedirectAttributes redirectAttributes,
+                              HttpServletResponse response) throws IOException{
+
+        if(!boardService.isMemberWriter((Long)session.getAttribute(SessionConstant.MEMBER_ID), boardId)) {
+           ScriptUtils.alertAndBackPage(response,"게시글 작성자가 아니라 게시글 삭제가 불가능합니다.");
+        } else {
+            boardService.deleteBoard(boardId);
+        }
+
+
         redirectAttributes.addAttribute("studyId", studyId);
         return "redirect:/study/{studyId}";
     }
@@ -136,14 +152,31 @@ public class BoardController {
     // 게시글의 좋아요 클릭 시 작업
     @PostMapping("/like")
     @ResponseBody
-    public List<Object> boardLikeProcess(@RequestParam(name="board_id") Long boardId,
-                              @RequestParam(name="member_id") Long memberId) {
-        LikeOnBoard likeOnBoard = likeOnBoardService.updateLike(boardId, memberId);
-        Integer likeCnt = likeOnBoardService.countLikeCnt(boardId, memberId);
-        List<Object> data = new ArrayList<>();
-        data.add(String.valueOf(likeOnBoard.getLikeCheck()));
-        data.add(likeCnt);
-        return data;
+    public Map<String,String> boardLikeProcess(@RequestParam(name="board_id") Long boardId,
+                                               @RequestParam(name="member_id") Long memberId,
+                                               HttpSession session) {
+
+        Map<String, String> map = new HashMap<>();
+
+        // 로그인 여부 확인 로직
+        if (memberId == null || !memberId.equals((Long)session.getAttribute(SessionConstant.MEMBER_ID))) {
+            map.put("notLoginMember", "로그인이 필요한 작업입니다. 로그인 페이지로 이동하시겠습니까?");
+        } else if ((boardRepository.findById(boardId).orElse(null) == null)) { // 게시글 존재 여부 확인 로직
+            map.put("notExistBoard","해당 게시글이 존재하지 않습니다.");
+        } else {
+            if (!likeOnBoardService.isExistingLikeOnBoard(boardId, memberId)) {
+                // LikeOnBoard 새로 생성
+                likeOnBoardService.getLikeOnBoard(boardId, memberId);
+            }
+
+            
+            LikeOnBoard likeOnBoard = likeOnBoardService.updateLike(boardId, memberId);
+            Integer likeCnt = likeOnBoardService.countLikeCnt(boardId, memberId);
+            map.put("likeCheck",String.valueOf(likeOnBoard.getLikeCheck()));
+            map.put("likeCnt", String.valueOf(likeCnt));
+        }
+
+        return map;
     }
